+++ lib/eligible/v1_0/estimate_service_line.rb
@@ -5,5 +5,13 @@ module Eligible
     class EstimateServiceLine < RestAPIBase
       ENDPOINT_NAME = 'estimate_service_lines'.freeze
     end
+
+    def self.update(_params, _opts = {})
+      fail NotImplementedError, "Not an allowed operation for this endpoint"
+    end
+
+    def self.delete(_params, _opts = {})
+      fail NotImplementedError, "Not an allowed operation for this endpoint"
+    end
   end
 end