+++ lib/eligible/v1_0/payment_report.rb
@@ -2,8 +2,8 @@
 
 module Eligible
   module V1_0
-    class InsuranceCompany < RestAPIBase
-      ENDPOINT_NAME = 'insurance_companies'.freeze
+    class PaymentReport < RestAPIBase
+      ENDPOINT_NAME = 'payment_reports'.freeze
 
       def self.create(_params, _opts = {})
         fail NotImplementedError, "Not an allowed operation for this endpoint"