+++ lib/eligible/v1_0/session.rb
@@ -2,18 +2,18 @@
 
 module Eligible
   module V1_0
-    class InsuranceCompany < RestAPIBase
-      ENDPOINT_NAME = 'insurance_companies'.freeze
-
-      def self.create(_params, _opts = {})
-        fail NotImplementedError, "Not an allowed operation for this endpoint"
-      end
+    class Session < RestAPIBase
+      ENDPOINT_NAME = 'sessions'.freeze
 
       def self.update(_params, _opts = {})
         fail NotImplementedError, "Not an allowed operation for this endpoint"
       end
 
       def self.delete(_params, _opts = {})
+        fail NotImplementedError, "Not an allowed operation for this endpoint"
+      end
+
+      def self.list(_params, _opts = {})
         fail NotImplementedError, "Not an allowed operation for this endpoint"
       end
     end