+++ lib/eligible/v1_0/verification.rb
@@ -2,12 +2,8 @@
 
 module Eligible
   module V1_0
-    class InsuranceCompany < RestAPIBase
-      ENDPOINT_NAME = 'insurance_companies'.freeze
-
-      def self.create(_params, _opts = {})
-        fail NotImplementedError, "Not an allowed operation for this endpoint"
-      end
+    class Verification < RestAPIBase
+      ENDPOINT_NAME = 'verifications'.freeze
 
       def self.update(_params, _opts = {})
         fail NotImplementedError, "Not an allowed operation for this endpoint"