+++ spec/lib/v1_0/estimate_service_line_spec.rb
@@ -4,7 +4,6 @@ describe 'Eligible::V1_0::EstimateServiceLine' do
   let(:response) { { success: true } }
   before(:each) do
     allow(Eligible::Util).to receive(:convert_to_eligible_object).with(response, api_key).and_return('success')
-    allow(Eligible::V1_0::EstimateServiceLine).to receive(:endpoint_name).and_return('estimate_service_lines')
   end
 
   describe '.retrieve' do