+++ spec/lib/v1_0/patient_statement_spec.rb
@@ -4,7 +4,6 @@ describe 'Eligible::V1_0::PatientStatement' do
   let(:response) { { success: true } }
   before(:each) do
     allow(Eligible::Util).to receive(:convert_to_eligible_object).with(response, api_key).and_return('success')
-    allow(Eligible::V1_0::PatientStatement).to receive(:endpoint_name).and_return('patient_statements')
   end
 
   describe '.retrieve' do
@@ -46,21 +45,14 @@ describe 'Eligible::V1_0::PatientStatement' do
   end
 
   describe "patient_statement events" do
-
-    before(:each) do
-      allow(Eligible::V1_0::PatientStatement).to receive(:statement_id).with(params).and_return('pst_9bcb7c733e0242439575a299')
-    end
-
     context '.finalize' do
       it 'should call Eligible.request with proper url' do
         params[:id] = 'pst_9bcb7c733e0242439575a299'
-        allow(Eligible::V1_0::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
         allow(Eligible).to receive(:request).with(:post, '/patient_statements/pst_9bcb7c733e0242439575a299/finalize', api_key, params, {}).and_return([response, api_key])
         expect(Eligible::V1_0::PatientStatement.finalize(params[:id], api_key: api_key)).to eq 'success'
       end
 
       it 'should raise error if patient_statement id is not present' do
-        allow(Eligible::V1_0::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
         expect { Eligible::V1_0::PatientStatement.finalize(nil, api_key: api_key) }.to raise_error(ArgumentError)
       end
     end
@@ -116,13 +108,11 @@ describe 'Eligible::V1_0::PatientStatement' do
     context '.send' do
       it 'should call Eligible.request with proper url' do
         params[:id] = 'pst_9bcb7c733e0242439575a299'
-        allow(Eligible::V1_0::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
         allow(Eligible).to receive(:request).with(:post, '/patient_statements/pst_9bcb7c733e0242439575a299/send', api_key, params, {}).and_return([response, api_key])
         expect(Eligible::V1_0::PatientStatement.send(params[:id], api_key: api_key)).to eq 'success'
       end
 
       it 'should raise error if patient_statement id is not present' do
-        allow(Eligible::V1_0::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
         expect { Eligible::V1_0::PatientStatement.send(nil, api_key: api_key) }.to raise_error(ArgumentError)
       end
     end
@@ -130,13 +120,11 @@ describe 'Eligible::V1_0::PatientStatement' do
     context '.void' do
       it 'should call Eligible.request with proper url' do
         params[:id] = 'pst_9bcb7c733e0242439575a299'
-        allow(Eligible::V1_0::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
         allow(Eligible).to receive(:request).with(:post, '/patient_statements/pst_9bcb7c733e0242439575a299/void', api_key, params, {}).and_return([response, api_key])
         expect(Eligible::V1_0::PatientStatement.void(params[:id], api_key: api_key)).to eq 'success'
       end
 
       it 'should raise error if patient_statement id is not present' do
-        allow(Eligible::V1_0::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
         expect { Eligible::V1_0::PatientStatement.void(nil, api_key: api_key) }.to raise_error(ArgumentError)
       end
     end
@@ -144,13 +132,11 @@ describe 'Eligible::V1_0::PatientStatement' do
     context '.mark_uncollectible' do
       it 'should call Eligible.request with proper url' do
         params[:id] = 'pst_9bcb7c733e0242439575a299'
-        allow(Eligible::V1_0::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
         allow(Eligible).to receive(:request).with(:post, '/patient_statements/pst_9bcb7c733e0242439575a299/mark_uncollectible', api_key, params, {}).and_return([response, api_key])
         expect(Eligible::V1_0::PatientStatement.mark_uncollectible(params[:id], api_key: api_key)).to eq 'success'
       end
 
       it 'should raise error if patient_statement id is not present' do
-        allow(Eligible::V1_0::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
         expect { Eligible::V1_0::PatientStatement.mark_uncollectible(nil, api_key: api_key) }.to raise_error(ArgumentError)
       end
     end
@@ -158,13 +144,11 @@ describe 'Eligible::V1_0::PatientStatement' do
     context '.reestimate' do
       it 'should call Eligible.request with proper url' do
         params[:id] = 'pst_9bcb7c733e0242439575a299'
-        allow(Eligible::V1_0::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
         allow(Eligible).to receive(:request).with(:post, '/patient_statements/pst_9bcb7c733e0242439575a299/reestimate', api_key, params, {}).and_return([response, api_key])
         expect(Eligible::V1_0::PatientStatement.reestimate(params, api_key: api_key)).to eq 'success'
       end
 
       it 'should raise error if patient_statement id is not present' do
-        allow(Eligible::V1_0::PatientStatement).to receive(:statement_id).with(params[:id]).and_return('pst_9bcb7c733e0242439575a299')
         expect { Eligible::V1_0::PatientStatement.reestimate(params, api_key: api_key) }.to raise_error(ArgumentError)
       end
     end